Add logging in most of the program

+++ main.py
@@ -13,10 +13,7 @@ from functions import angles_generator, angles_random, vec_to_dist
 from askInterEvol import runAlign
 import filedownload as fd
 from alignInterolog import runProfit
-
-FORMAT = '%(asctime)s - %(name)s - %(levelname)8s : %(message)s'
-logging.basicConfig(format=FORMAT,level=logging.DEBUG)
-logger = logging.getLogger('main')
+import runMini as mini
 
 parser = argparse.ArgumentParser()
 parser.add_argument('-rec',required=True,help='Recepter\'s file path')
@@ -24,7 +21,6 @@ parser.add_argument('-lig',required=True,help='Ligand\'s file path')
 parser.add_argument('-n',help='Number of sampling')
 parser.add_argument('-o',default='out',help='The directory where program\'s output sotres')
 
-logger.debug('Start parsing arguments')
 args = parser.parse_args()
 
 
@@ -41,6 +37,20 @@ for d,n in fld_lst:
     t = os.path.join(FLD['OUT'],n)
     pathlib.Path(t).mkdir(parents=True,exist_ok=True)
     FLD[d] = t
+
+# Define logger
+FORMAT = '%(asctime)s - %(name)s - %(levelname)8s : %(message)s'
+logger = logging.getLogger('main')
+logger.setLevel(logging.DEBUG)
+fh = logging.FileHandler(os.path.join(FLD['OUT'],'log.txt'))
+fh.setLevel(logging.INFO)
+ch = logging.StreamHandler()
+ch.setLevel(logging.DEBUG)
+formatter = logging.Formatter(FORMAT)
+fh.setFormatter(formatter)
+ch.setFormatter(formatter)
+logger.addHandler(fh)
+logger.addHandler(ch)
 
 LIG = re.match(PDB,pathlib.PurePath(args.lig).name).group(1)
 REC = re.match(PDB,pathlib.PurePath(args.rec).name).group(1)
@@ -64,6 +74,7 @@ ligf = os.path.join(FLD['FASTA'],LIG+'.fasta')
 recf = os.path.join(FLD['FASTA'],REC+'.fasta')
 lig.write_fasta(ligf)
 rec.write_fasta(recf)
+logger.debug("Start download. Please wait...")
 dico = runAlign(recf,ligf,FLD['INTER'])
 
 
@@ -71,17 +82,12 @@ dico = runAlign(recf,ligf,FLD['INTER'])
 ###  Alignment with Profit   ###
 ################################
 
-
+logger.debug("Start of alignment with ProFit")
 for key in dico.keys():
-	#print(key)
-    #print(dico[key])
     liste = dico[key]
-    #print(liste[3],liste[2])
     deg = min(int(liste[0][:-1]),int(liste[1][:-1]))
     runProfit(lig.path, rec.path, os.path.join(FLD['INTER'],key+".pdb"), liste[3], liste[2])
-
-
-
+logger.debug("End of alignment with ProFit")
 
 lig_aligned = pdb.Protein.from_pdb_file(os.path.join(FLD['PRO'],LIG+'_aligned.pdb'))
 lig_aligned.name = LIG
@@ -92,7 +98,9 @@ cpx = Complex(rec,lig_aligned)
 ################################
 
 
+logger.debug('Start sampling')
 for idx,l in enumerate(angles_generator(n_samples,deg=deg)):
+    logger.info('Ligand {:>6}\'s rotatation No. {:05d}({})'.format(cpx.lig.name,idx,l))
     A = cpx.rotations(l[0],l[1],l[2],l[3],l[4])
     D = cpx.ca_dist(A)
     i,j = np.unravel_index(D.argmin(), D.shape)
@@ -100,10 +108,14 @@ for idx,l in enumerate(angles_generator(n_samples,deg=deg)):
     if m <5:
         A = A+vec_to_dist(cpx.rec.get_ca()[i],A[cpx.lig.get_ca_ind()][j],5)
     cpx.lig.write_atoms(os.path.join(FLD['PRO'],'B{:05d}.pdb'.format(idx)),A)
-    
+logger.debug("End sampling")
 
+logger.debug('Start of minimizer')
 for k in range(n_samples):
-    subprocess.call(["python3", 'Minimizer/runMini.py', '-rec',cpx.rec.path, '-lig',os.path.join(FLD['PRO'],'B{:05d}.pdb'.format(k)),'-o',FLD['OUT']])
-    out_file = os.path.join(FLD['OUT'],'pdb_mini/{}_B{:05d}_min1.pdb'.format(cpx.rec.name,k))
-    cpx_out_file = os.path.join(FLD['CPX'],'{}_B{:05d}.pdb'.format(cpx.rec.name,k))
-    subprocess.call("sed '1d' {} > tmp.txt; cat {} tmp.txt > {}; rm tmp.txt".format(out_file,cpx.rec.path,cpx_out_file),shell=True)
+    successed = mini.run(cpx.rec.path,os.path.join(FLD['PRO'],'B{:05d}.pdb'.format(k)),FLD['OUT'],k)
+    if successed:
+        out_file = os.path.join(FLD['OUT'],'pdb_mini/{}_B{:05d}_min_{:05d}.pdb'.format(cpx.rec.name,k,k))
+        cpx_out_file = os.path.join(FLD['CPX'],'{}_B{:05d}.pdb'.format(cpx.rec.name,k))
+        subprocess.call("sed '1d' {} > tmp.txt; cat {} tmp.txt > {}; rm tmp.txt".format(out_file,cpx.rec.path,cpx_out_file),shell=True)
+logger.debug('End of minimizer')
+logger.debug('End of program')
